refactor: split joinus button component
